ref(installer): mirrorlist and pacman conf writer moves to pm/mirrors.py
* next to MirrorListHandler, which produced the region list it writes
